refactor: コメント追記、StateProvider の値更新記法修正、画面とモデルの判別ができるようなファイル名に修正、それに伴うファイルおよびフォルダの分割
