+++ lib/presentation/diagnosis/depression/depression_type_table.dart
