+++ lib/presentation/diagnosis/manic/manic_type_table_page.dart
@@ -1,59 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mood_trend_flutter/presentation/diagnosis/depression_type_diagnosis.dart';
-import 'package:mood_trend_flutter/presentation/diagnosis/manic_type_diagnosis.dart';
 import 'package:mood_trend_flutter/utils/app_colors.dart';
 import 'package:mood_trend_flutter/utils/page_navigator.dart';
 
-abstract class ManicWorkSheet {
-  late final String plus_5;
-  late final String plus_4;
-  late final String plus_3;
-  late final String plus_2;
-  late final String plus_1;
-}
+import 'entity/manic_worksheet.dart';
+import 'manic_type_diagnosis_page.dart';
 
-class IdeaType extends ManicWorkSheet {
-  @override
-  String get plus_5 => 'アイデアが成功すると確信し、実現に向けて実際に動く';
-  @override
-  String get plus_4 => '次々と湧き上がるアイデアを具現化しようと奮闘する';
-  @override
-  String get plus_3 => 'アイデアが湧き上がり、絶え間なく考え続ける';
-  @override
-  String get plus_2 => 'アイデアを簡単に思いつくことができる';
-  @override
-  String get plus_1 => 'アイデアを考えようと思えば可能である';
-}
-
-class ElationType extends ManicWorkSheet {
-  @override
-  String get plus_5 => '自分は絶対に正しく、すべてが上手くいくと確信する';
-  @override
-  String get plus_4 => 'どんなことでも自分は成功すると確信している';
-  @override
-  String get plus_3 => 'どんな状況でも上手くいく感覚が強い';
-  @override
-  String get plus_2 => '前向きで挑戦的な気持ちがあり、ポジティブなエネルギーを感じる';
-  @override
-  String get plus_1 => '普段より前向きな気分である';
-}
-
-class ActivityType extends ManicWorkSheet {
-  @override
-  String get plus_5 => '何日も1日中活動を続けているが、疲れを感じずに精力的に行動している';
-  @override
-  String get plus_4 => '1日が予定でいっぱいである';
-  @override
-  String get plus_3 => '積極的でエネルギッシュな気分で、予定をたくさん詰める';
-  @override
-  String get plus_2 => '普段よりも活発に動く気分である';
-  @override
-  String get plus_1 => '活動が可能である';
-}
-
-class ManicTypeTable extends ConsumerWidget {
-  const ManicTypeTable({super.key, required this.manicType});
+/// 躁のタイプを表示するテーブル画面
+class ManicTypeTablePage extends ConsumerWidget {
+  const ManicTypeTablePage({super.key, required this.manicType});
   final ManicType manicType;
 
   @override
@@ -155,7 +110,7 @@ class ManicTypeTable extends ConsumerWidget {
                 fixedSize: const Size(330, 60),
               ),
               onPressed: () {
-                PageNavigator.push(context, const DepressionTypeDignosis());
+                PageNavigator.push(context, const DepressionTypeDignosisPage());
               },
               child: const Text(
                 '同意して次へ',
